Refactor Button definition
